requirements.txt
# Brand Guardian - Video Compliance Audit Pipeline
# Core API
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.10.0

# Video Download
yt-dlp>=2024.12.0

# Speech-to-Text
openai-whisper>=20240930

# OCR
easyocr>=1.7.2

# Embeddings & Vector Store
sentence-transformers>=3.3.0
faiss-cpu>=1.9.0

# PDF Processing
pypdf>=5.1.0

# HTTP (for Ollama API)
requests>=2.32.0

# Utilities
python-multipart>=0.0.12
